+++ app/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+- 2017-29-12 | v.0.2.7:
+
+  Added charts with Volume, Price, Marketcap
+
 - 2017-26-12 | v.0.2.6:
 
   Added Project Detailed Page